Move some fns to key-utils, fix gcp secret names
