+++ typescript/infra/src/agents/gcp.ts
@@ -7,7 +7,6 @@ import { fetchGCPSecret, setGCPSecret } from '../utils/gcloud';
 import { execCmd, include } from '../utils/utils';
 
 import { AgentKey, isValidatorKey, keyIdentifier } from './agent';
-import { getAllKeys, getKey } from './key-utils';
 import { KEY_ROLE_ENUM } from './roles';
 
 // This is the type for how the keys are persisted in GCP
@@ -19,11 +18,6 @@ export interface SecretManagerPersistedKeys {
   // Exists if key is an attestation key
   // TODO: Add this to the type
   chainName?: string;
-}
-
-interface KeyAsAddress {
-  identifier: string;
-  address: string;
 }
 
 interface UnfetchedKey {
@@ -153,105 +147,4 @@ export class AgentGCPKey<
       address,
     };
   }
-}
-
-export async function deleteAgentKeys<Networks extends ChainName>(
-  agentConfig: AgentConfig<Networks>,
-) {
-  const keys = getAllKeys(agentConfig);
-  await Promise.all(keys.map((key) => key.delete()));
-  await execCmd(
-    `gcloud secrets delete ${addressesIdentifier(
-      agentConfig.environment,
-    )} --quiet`,
-  );
-}
-
-export async function createAgentKeysIfNotExists<Networks extends ChainName>(
-  agentConfig: AgentConfig<Networks>,
-) {
-  const keys = getAllKeys(agentConfig);
-
-  await Promise.all(
-    keys.map(async (key) => {
-      return key.createIfNotExists();
-    }),
-  );
-
-  await persistAddresses(
-    agentConfig.environment,
-    keys.map((key) => key.serializeAsAddress()),
-  );
-}
-
-export async function rotateKey<Networks extends ChainName>(
-  agentConfig: AgentConfig<Networks>,
-  role: KEY_ROLE_ENUM,
-  chainName: Networks,
-) {
-  const key = getKey(agentConfig, role, chainName);
-  await key.update();
-  const keyIdentifier = key.identifier;
-  const addresses = await fetchGCPKeyAddresses(agentConfig.environment);
-  const filteredAddresses = addresses.filter((_) => {
-    return _.identifier !== keyIdentifier;
-  });
-
-  filteredAddresses.push(key.serializeAsAddress());
-  await persistAddresses(agentConfig.environment, filteredAddresses);
-}
-
-async function persistAddresses(environment: string, keys: KeyAsAddress[]) {
-  await setGCPSecret(addressesIdentifier(environment), JSON.stringify(keys), {
-    environment: environment,
-  });
-}
-
-// This function returns all the GCP keys for a given outbox chain in a dictionary where the key is the identifier
-export async function fetchAgentGCPKeys<Networks extends ChainName>(
-  _agentConfig: AgentConfig<Networks>,
-  _chainName: Networks,
-  _validatorCount: number,
-): Promise<Record<string, AgentGCPKey<Networks>>> {
-  // const environment = agentConfig.environment;
-  // const secrets = await Promise.all(
-  //   KEY_ROLES.map(async (role) => {
-  //     if (role === KEY_ROLE_ENUM.Validator) {
-  //       return Promise.all(
-  //         [...Array(validatorCount).keys()].map(async (index) => {
-  //           const key = new AgentGCPKey(environment, role, chainName, index);
-  //           await key.fetch();
-  //           return [key.identifier, key];
-  //         }),
-  //       );
-  //     } else if (role === KEY_ROLE_ENUM.Relayer) {
-  //       return Promise.all(
-  //         agentConfig
-  //           .domainNames
-  //           .filter((d) => d !== chainName)
-  //           .map(async (remote) => {
-  //             const key = new AgentGCPKey(environment, role, chainName, remote);
-  //             await key.fetch();
-  //             return [key.identifier, key];
-  //           })
-  //       );
-  //     } else {
-  //       const key = new AgentGCPKey(environment, role, chainName);
-  //       await key.fetch();
-  //       return [[key.identifier, key]];
-  //     }
-  //   }),
-  // );
-
-  // return Object.fromEntries(secrets.flat(1));
-  return {};
-}
-
-async function fetchGCPKeyAddresses(environment: string) {
-  const addresses = await fetchGCPSecret(addressesIdentifier(environment));
-  return addresses as KeyAsAddress[];
-}
-
-function addressesIdentifier(environment: string) {
-  return `abacus-${environment}-key-addresses`;
 }
